Build system upgrade, move everything to use rust cargo instead of crates Rust cargo for libpal and for exercises set will be located per set There still needs to be solve how will I build exercies for a local test and not a part of a cargo for development and testing
