+++ libpal/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "libpal"
+version = "0.1.1"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
