+++ set1/solutions/src/exercise3.rs
@@ -1,6 +1,5 @@
 #![allow(non_snake_case)]
-extern crate base64Tools;
-extern crate libpal;
+use libpal::pal as libpal;
 
 fn CrackXorTest()
 {
